+++ Sources/Internal/GrpcClient.swift
@@ -58,6 +58,7 @@ enum GrpcClientRequestHeaders {
   static let googApiClient = "x-goog-api-client"
   static let clientPlatform = "x-client-platform"
   static let clientVersion = "x-client-version"
+  static let sqlConnectAffinity = "x-firebase-sqlconnect-affinity"
 }
 
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
@@ -113,6 +114,7 @@ actor DataConnectGrpcClient: GrpcClient, CustomStringConvertible {
     let connectorName =
       "projects/\(projectId)/locations/\(connectorConfig.location)/services/\(connectorConfig.serviceId)/connectors/\(connectorConfig.connector)"
     let googRequestHeaderValue = "location=\(connectorConfig.location)&frontend=data"
+    let sqlConnectAffinityHeaderValue = "\(projectId)\(connectorConfig.serviceId)"
 
     let header =
       "gl-swift/\(Version.swiftVersion()) fire/\(Version.sdkVersion) \(Version.platformVersionHeader()) grpc-swift/"
@@ -146,7 +148,8 @@ actor DataConnectGrpcClient: GrpcClient, CustomStringConvertible {
       appCheck: appCheck,
       callerSDKType: callerSDKType,
       googRequestHeaderValue: googRequestHeaderValue,
-      googApiClientHeaderValue: googApiClientHeaderValue
+      googApiClientHeaderValue: googApiClientHeaderValue,
+      sqlConnectAffinityHeaderValue: sqlConnectAffinityHeaderValue
     )
 
     streamingClient = StreamingGrpcClient(
@@ -157,7 +160,8 @@ actor DataConnectGrpcClient: GrpcClient, CustomStringConvertible {
       appCheck: appCheck,
       callerSDKType: callerSDKType,
       googRequestHeaderValue: googRequestHeaderValue,
-      googApiClientHeaderValue: googApiClientHeaderValue
+      googApiClientHeaderValue: googApiClientHeaderValue,
+      sqlConnectAffinityHeaderValue: sqlConnectAffinityHeaderValue
     )
   }
 
@@ -276,7 +280,8 @@ actor DataConnectGrpcClient: GrpcClient, CustomStringConvertible {
                                 auth: Auth,
                                 appCheck: AppCheckInterop?,
                                 googRequestHeaderValue: String,
-                                googApiClientHeaderValue: String) async -> CallOptions {
+                                googApiClientHeaderValue: String,
+                                sqlConnectAffinityHeaderValue: String) async -> CallOptions {
     var headers = HPACKHeaders()
 
     if app.isDataCollectionDefaultEnabled {
@@ -288,6 +293,10 @@ actor DataConnectGrpcClient: GrpcClient, CustomStringConvertible {
       headers.add(name: GrpcClientRequestHeaders.googApiClient, value: googApiClientHeaderValue)
       headers.add(name: GrpcClientRequestHeaders.clientPlatform, value: "ios")
       headers.add(name: GrpcClientRequestHeaders.clientVersion, value: Version.sdkVersion)
+      headers.add(
+        name: GrpcClientRequestHeaders.sqlConnectAffinity,
+        value: sqlConnectAffinityHeaderValue
+      )
     }
 
     // Add Auth token if available