+++ Sources/Internal/UnaryGrpcClient.swift
@@ -30,6 +30,7 @@ actor UnaryGrpcClient: GrpcClient {
   private let callerSDKType: CallerSDKType
   private let googRequestHeaderValue: String
   private let googApiClientHeaderValue: String
+  private let sqlConnectAffinityHeaderValue: String
 
   private lazy var unaryClient: FirebaseDataConnectUnaryClient? = {
     do {
@@ -53,7 +54,8 @@ actor UnaryGrpcClient: GrpcClient {
        appCheck: AppCheckInterop?,
        callerSDKType: CallerSDKType,
        googRequestHeaderValue: String,
-       googApiClientHeaderValue: String) {
+       googApiClientHeaderValue: String,
+       sqlConnectAffinityHeaderValue: String) {
     self.app = app
     self.connectorName = connectorName
     self.serverSettings = serverSettings
@@ -62,6 +64,7 @@ actor UnaryGrpcClient: GrpcClient {
     self.callerSDKType = callerSDKType
     self.googRequestHeaderValue = googRequestHeaderValue
     self.googApiClientHeaderValue = googApiClientHeaderValue
+    self.sqlConnectAffinityHeaderValue = sqlConnectAffinityHeaderValue
   }
 
   func executeQuery<
@@ -267,7 +270,8 @@ actor UnaryGrpcClient: GrpcClient {
       auth: auth,
       appCheck: appCheck,
       googRequestHeaderValue: googRequestHeaderValue,
-      googApiClientHeaderValue: googApiClientHeaderValue
+      googApiClientHeaderValue: googApiClientHeaderValue,
+      sqlConnectAffinityHeaderValue: sqlConnectAffinityHeaderValue
     )
   }
 }